Add directors to series add method & minor changes on css

+++ backend/app/Http/Controllers/MovieController.php
@@ -29,7 +29,7 @@ class MovieController extends Controller
             'user_id' => $user->id,
             'content' => request()->content,
             'rating' => request()->rating,
-            'release_year' => (int) request()->year
+            'release_year' => request()->year
         ]);
 
         /*handle img*/
